using uint for count of strategy usages

+++ SimpleGameSolver/BrownMethodBase.cs
@@ -7,9 +7,9 @@ namespace SimpleGameSolver
 {
     public abstract class BrownMethodBase
     {
-        public abstract List<List<List<int>>> Solve(double[,] fpp, double[,] spp, List<int> fpsp, List<int> spsp, int iterationsCount);
+        public abstract List<List<List<ulong>>> Solve(double[,] fpp, double[,] spp, List<ulong> fpsp, List<ulong> spsp, int iterationsCount);
 
-        protected int Choice1(double[,] A, List<int> otherPlayerStartegy, int last)
+        protected int Choice1(double[,] A, List<ulong> otherPlayerStartegy, int last)
         {
             int result = 0;
             List<double> tmp = new List<double>();//список выйгрышей при разном выборе стратегии
@@ -27,7 +27,7 @@ namespace SimpleGameSolver
             return result;
         }
 
-        protected int Choice2(double[,] A, List<int> otherPlayerStartegy, int last)
+        protected int Choice2(double[,] A, List<ulong> otherPlayerStartegy, int last)
         {
             int result = 0;
             List<double> tmp = new List<double>();//список выйгрышей при разном выборе стратегии
